chore(bot): updated command permissions and command cleanup
