+++ src/commands/testing/test-join.ts
@@ -1,10 +1,11 @@
-import { PermissionsBitField, SlashCommandBuilder } from 'discord.js';
+import { PermissionFlagsBits, SlashCommandBuilder } from 'discord.js';
 
 import { Command } from '@/types/CommandTypes.js';
 
 const command: Command = {
   data: new SlashCommandBuilder()
-    .setName('testjoin')
+    .setName('test-join')
+    .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
     .setDescription('Simulates a new member joining'),
 
   execute: async (interaction) => {
@@ -12,17 +13,6 @@ const command: Command = {
     const guild = interaction.guild;
 
     await interaction.deferReply({ flags: ['Ephemeral'] });
-
-    if (
-      !interaction.memberPermissions!.has(
-        PermissionsBitField.Flags.Administrator,
-      )
-    ) {
-      await interaction.editReply({
-        content: 'You do not have permission to use this command.',
-      });
-      return;
-    }
 
     const fakeMember = await guild.members.fetch(interaction.user.id);
     guild.client.emit('guildMemberAdd', fakeMember);
