+++ src/commands/util/restart.ts
@@ -1,9 +1,8 @@
-import { PermissionsBitField, SlashCommandBuilder } from 'discord.js';
+import { PermissionFlagsBits, SlashCommandBuilder } from 'discord.js';
 import { exec } from 'child_process';
 import { promisify } from 'util';
 
 import { Command } from '@/types/CommandTypes.js';
-import { loadConfig } from '@/util/configLoader.js';
 import {
   NotificationType,
   notifyManagers,
@@ -16,32 +15,12 @@ const execAsync = promisify(exec);
 const command: Command = {
   data: new SlashCommandBuilder()
     .setName('restart')
+    .setDefaultMemberPermissions(PermissionFlagsBits.ManageGuild)
     .setDescription('(Manager Only) Restart the bot'),
   execute: async (interaction) => {
     if (!interaction.isChatInputCommand() || !interaction.guild) return;
 
     await interaction.deferReply({ flags: ['Ephemeral'] });
-
-    const config = loadConfig();
-    const managerRoleId = config.roles.staffRoles.find(
-      (role) => role.name === 'Manager',
-    )?.roleId;
-
-    const member = await interaction.guild.members.fetch(interaction.user.id);
-    const hasManagerRole = member?.roles.cache.has(managerRoleId || '');
-
-    if (
-      !hasManagerRole &&
-      !interaction.memberPermissions?.has(
-        PermissionsBitField.Flags.Administrator,
-      )
-    ) {
-      await interaction.editReply({
-        content:
-          'You do not have permission to restart the bot. This command is restricted to users with the Manager role.',
-      });
-      return;
-    }
 
     await interaction.editReply({
       content: 'Restarting the bot... This may take a few moments.',
